// src/modules/horizon/proyeccion/presupuesto/types/ProyeccionData.ts
export interface ProyeccionMensualData {
  nominas: number[];
  autonomos: number[];
  alquiler: number[];
  intereses: number[];
  otrosIngresos: number[];
  totalIngresos: number[];
  gastosAlquileres: number[];
  gastosPersonales: number[];
  gastosAutonomo: number[];
  irpf: number[];
  totalGastos: number[];
  hipotecas: number[];
  prestamos: number[];
  totalFinanciacion: number[];
  flujoCaja: number[];
  cajaInicial: number[];
  cajaFinal: number[];
}